+++ R/cnd-message.R
@@ -151,7 +151,7 @@ cnd_header <- function(cnd, ...) {
   if (is_null(cnd[["header"]])) {
     UseMethod("cnd_header")
   } else {
-    exec_cnd_method("header", cnd)
+    exec_cnd_method("header", cnd, ...)
   }
 }
 #' @export
@@ -165,7 +165,7 @@ cnd_body <- function(cnd, ...) {
   if (is_null(cnd[["body"]])) {
     UseMethod("cnd_body")
   } else {
-    exec_cnd_method("body", cnd)
+    exec_cnd_method("body", cnd, ...)
   }
 }
 #' @export
@@ -179,7 +179,7 @@ cnd_footer <- function(cnd, ...) {
   if (is_null(cnd[["footer"]])) {
     UseMethod("cnd_footer")
   } else {
-    exec_cnd_method("footer", cnd)
+    exec_cnd_method("footer", cnd, ...)
   }
 }
 #' @export
@@ -187,14 +187,14 @@ cnd_footer.default <- function(cnd, ...) {
   chr()
 }
 
-exec_cnd_method <- function(name, cnd) {
+exec_cnd_method <- function(name, cnd, ...) {
   method <- cnd[[name]]
 
   if (is_function(method)) {
-    method(cnd)
+    method(cnd, ...)
   } else if (is_bare_formula(method)) {
     method <- as_function(method)
-    method(cnd)
+    method(cnd, ...)
   } else if (is_character(method)) {
     method
   } else {
@@ -224,7 +224,12 @@ cnd_message_format_prefixed <- function(cnd,
     prefix <- col_yellow(capitalise(type))
   }
 
-  call <- format_error_call(cnd$call)
+  evalq({
+    if (is_true(peek_option("rlang:::error_highlight"))) {
+      local_error_highlight()
+    }
+    call <- format_error_call(cnd$call)
+  })
 
   message <- cnd_message_format(cnd, ..., alert = alert)
   message <- strip_trailing_newline(message)