+++ tests/testthat/test-arg.R
@@ -310,3 +310,14 @@ test_that("arg_match() mentions correct call if wrong type is supplied (#1388)",
     (expect_error(g(1)))
   })
 })
+
+test_that("arg_match() backtrace highlights call and arg", {
+  f <- function(x) g(x)
+  g <- function(x) h(x)
+  h <- function(my_arg = c("foo", "bar")) arg_match(my_arg)
+  err <- catch_error(f("f"))
+
+  expect_snapshot({
+    print_highlighted_trace(err)
+  })
+})